Initial commit with refactored code
