+++ .gitignore
@@ -1,4 +1,5 @@
 /venv/
+/env/
 .idea/
 .vscode/
 __pycache__/
