+++ requirements.txt
@@ -1,2 +1,3 @@
-Flask==0.12.3
+werkzeug=0.15.6
+Flask==1.1.1
 boto3==1.9.193
