+++ server/dynamodb/ddb_service.py
@@ -1,16 +1,15 @@
 import boto3
-import uuid
-import json
-import hashlib
 
 DYNAMO_ENDPOINT = 'http://localhost:8000'
 
 
 def create_resource():
     return boto3.resource('dynamodb', endpoint_url=DYNAMO_ENDPOINT, region_name='us-west-2')
-    
+
+
 def create_client():
     return boto3.client('dynamodb', endpoint_url=DYNAMO_ENDPOINT, region_name='us-west-2')
+
 
 def list_all_table_items():
     dynamodb = create_resource()
@@ -21,11 +20,7 @@ def list_all_table_items():
     return data
 
 
-
-
-
-
-def list_tables():    
+def list_tables():
     response = create_client().list_tables()
     table_list_detail = []
     table_list = response.get('TableNames')
@@ -34,16 +29,18 @@ def list_tables():
         table_detail = create_table_detail(table_info.get('Table'))
         table_list_detail.append(table_detail)
     return table_list_detail
-    
+
+
 def describe_table(table_name):
     table_detail = create_client().describe_table(TableName=table_name)
     return table_detail
 
+
 def create_table_detail(table):
     table_detail = {
-        'table_name' : table['TableName'],
-        'item_count' : table['ItemCount'],
-        'table_status' : table['TableStatus'],
-        'table_size_bytes' : table['TableSizeBytes']
+        'table_name': table['TableName'],
+        'item_count': table['ItemCount'],
+        'table_status': table['TableStatus'],
+        'table_size_bytes': table['TableSizeBytes']
     }
     return table_detail
