+++ start-app.py
@@ -3,11 +3,13 @@ import subprocess
 import threading
 import time
 
+
 def start_docker_images():
     print("Starting docker-compose")
     subprocess.call("docker-compose down")
     subprocess.call("docker-compose up")
-    
+
+
 def execute_docker_compose():   
     print("Starting Images: Localstack - DynamoDB")
     x = threading.Thread(target=start_docker_images, args = ())
@@ -16,13 +18,15 @@ def execute_docker_compose():
     print("Waiting for images finish starting ... ")
     time.sleep(25)
 
+
 def open_venv():
     print("Opening venv")
     # TODO: find a way to open a venv through python script
     # for now, do the process manually
     # windows:  venv\Scripts\activate
     # linux:    source env/bin/activate
-    
+
+
 def start_localstack_manager():
     print("Starting Localstack Manager")
     os.environ["FLASK_APP"] = "aws_flask.py"
@@ -33,7 +37,5 @@ def start_localstack_manager():
 
 if __name__ == '__main__':        
     #open_venv()
-    execute_docker_compose()
-    #start_localstack_manager()
-    
-    
+    #execute_docker_compose()
+    start_localstack_manager()
